Move template and schema to handler
