+++ cmd/authgear/server/wire_gen.go
@@ -16,7 +16,6 @@ import (
 	"github.com/authgear/authgear-server/pkg/auth/dependency/authenticator/recoverycode"
 	"github.com/authgear/authgear-server/pkg/auth/dependency/authenticator/totp"
 	"github.com/authgear/authgear-server/pkg/auth/dependency/challenge"
-	"github.com/authgear/authgear-server/pkg/auth/dependency/forgotpassword"
 	"github.com/authgear/authgear-server/pkg/auth/dependency/hook"
 	"github.com/authgear/authgear-server/pkg/auth/dependency/identity/anonymous"
 	"github.com/authgear/authgear-server/pkg/auth/dependency/identity/loginid"
@@ -32,7 +31,6 @@ import (
 	handler2 "github.com/authgear/authgear-server/pkg/auth/dependency/oidc/handler"
 	"github.com/authgear/authgear-server/pkg/auth/dependency/session"
 	redis3 "github.com/authgear/authgear-server/pkg/auth/dependency/session/redis"
-	"github.com/authgear/authgear-server/pkg/auth/dependency/sso"
 	"github.com/authgear/authgear-server/pkg/auth/dependency/user"
 	"github.com/authgear/authgear-server/pkg/auth/dependency/webapp"
 	"github.com/authgear/authgear-server/pkg/auth/dependency/welcomemessage"
@@ -1280,349 +1278,8 @@ func newWebAppRootHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppLoginHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	oAuthSSOConfig := identityConfig.OAuth
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	jwtKeyMaterials := deps.ProvideJWTKeyMaterials(secretConfig)
-	stateCodec := &sso.StateCodec{
-		AppID:       appID,
-		Clock:       clockClock,
-		Credentials: jwtKeyMaterials,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	messagingConfig := appConfig.Messaging
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	logger := hook.NewLogger(factory)
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	sessionConfig := appConfig.Session
-	cookieDef := session.NewSessionCookieDef(request, sessionConfig, serverConfig)
-	redisLogger := redis3.NewLogger(factory)
-	redisStore := &redis3.Store{
-		Redis:  redisHandle,
-		AppID:  appID,
-		Clock:  clockClock,
-		Logger: redisLogger,
-	}
-	eventStore := &redis2.EventStore{
-		Redis: redisHandle,
-		AppID: appID,
-	}
-	accessEventProvider := &auth.AccessEventProvider{
-		Store: eventStore,
-	}
-	sessionRand := _wireRandValue
-	sessionProvider := &session.Provider{
-		Request:      request,
-		Store:        redisStore,
-		AccessEvents: accessEventProvider,
-		ServerConfig: serverConfig,
-		Config:       sessionConfig,
-		Clock:        clockClock,
-		Random:       sessionRand,
-	}
-	userController := &flows.UserController{
-		Users:         userProvider,
-		SessionCookie: cookieDef,
-		Sessions:      sessionProvider,
-		Hooks:         hookProvider,
-	}
-	webAppFlow := &flows.WebAppFlow{
-		Config:         identityConfig,
-		Identities:     providerProvider,
-		Users:          userProvider,
-		Hooks:          hookProvider,
-		Interactions:   interactionProvider,
-		UserController: userController,
-	}
-	oAuthClientCredentials := deps.ProvideOAuthClientCredentials(secretConfig)
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	userInfoDecoder := sso.UserInfoDecoder{
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	oAuthProviderFactory := &sso.OAuthProviderFactory{
-		Endpoints:                endpointsProvider,
-		IdentityConfig:           identityConfig,
-		Credentials:              oAuthClientCredentials,
-		RedirectURL:              urlProvider,
-		Clock:                    clockClock,
-		UserInfoDecoder:          userInfoDecoder,
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	authenticateProviderImpl := &webapp.AuthenticateProviderImpl{
-		ServerConfig:         serverConfig,
-		SSOOAuthConfig:       oAuthSSOConfig,
-		ValidateProvider:     validateProviderImpl,
-		RenderProvider:       renderProviderImpl,
-		StateProvider:        stateProviderImpl,
-		SSOStateCodec:        stateCodec,
-		Interactions:         webAppFlow,
-		OAuthProviderFactory: oAuthProviderFactory,
-	}
 	loginHandler := &webapp2.LoginHandler{
-		Provider: authenticateProviderImpl,
 		Database: handle,
 	}
 	return loginHandler
@@ -1630,349 +1287,8 @@ func newWebAppLoginHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppSignupHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	oAuthSSOConfig := identityConfig.OAuth
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	jwtKeyMaterials := deps.ProvideJWTKeyMaterials(secretConfig)
-	stateCodec := &sso.StateCodec{
-		AppID:       appID,
-		Clock:       clockClock,
-		Credentials: jwtKeyMaterials,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	messagingConfig := appConfig.Messaging
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	logger := hook.NewLogger(factory)
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	sessionConfig := appConfig.Session
-	cookieDef := session.NewSessionCookieDef(request, sessionConfig, serverConfig)
-	redisLogger := redis3.NewLogger(factory)
-	redisStore := &redis3.Store{
-		Redis:  redisHandle,
-		AppID:  appID,
-		Clock:  clockClock,
-		Logger: redisLogger,
-	}
-	eventStore := &redis2.EventStore{
-		Redis: redisHandle,
-		AppID: appID,
-	}
-	accessEventProvider := &auth.AccessEventProvider{
-		Store: eventStore,
-	}
-	sessionRand := _wireRandValue
-	sessionProvider := &session.Provider{
-		Request:      request,
-		Store:        redisStore,
-		AccessEvents: accessEventProvider,
-		ServerConfig: serverConfig,
-		Config:       sessionConfig,
-		Clock:        clockClock,
-		Random:       sessionRand,
-	}
-	userController := &flows.UserController{
-		Users:         userProvider,
-		SessionCookie: cookieDef,
-		Sessions:      sessionProvider,
-		Hooks:         hookProvider,
-	}
-	webAppFlow := &flows.WebAppFlow{
-		Config:         identityConfig,
-		Identities:     providerProvider,
-		Users:          userProvider,
-		Hooks:          hookProvider,
-		Interactions:   interactionProvider,
-		UserController: userController,
-	}
-	oAuthClientCredentials := deps.ProvideOAuthClientCredentials(secretConfig)
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	userInfoDecoder := sso.UserInfoDecoder{
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	oAuthProviderFactory := &sso.OAuthProviderFactory{
-		Endpoints:                endpointsProvider,
-		IdentityConfig:           identityConfig,
-		Credentials:              oAuthClientCredentials,
-		RedirectURL:              urlProvider,
-		Clock:                    clockClock,
-		UserInfoDecoder:          userInfoDecoder,
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	authenticateProviderImpl := &webapp.AuthenticateProviderImpl{
-		ServerConfig:         serverConfig,
-		SSOOAuthConfig:       oAuthSSOConfig,
-		ValidateProvider:     validateProviderImpl,
-		RenderProvider:       renderProviderImpl,
-		StateProvider:        stateProviderImpl,
-		SSOStateCodec:        stateCodec,
-		Interactions:         webAppFlow,
-		OAuthProviderFactory: oAuthProviderFactory,
-	}
 	signupHandler := &webapp2.SignupHandler{
-		Provider: authenticateProviderImpl,
 		Database: handle,
 	}
 	return signupHandler
@@ -1980,349 +1296,8 @@ func newWebAppSignupHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppPromoteHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	oAuthSSOConfig := identityConfig.OAuth
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	jwtKeyMaterials := deps.ProvideJWTKeyMaterials(secretConfig)
-	stateCodec := &sso.StateCodec{
-		AppID:       appID,
-		Clock:       clockClock,
-		Credentials: jwtKeyMaterials,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	messagingConfig := appConfig.Messaging
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	logger := hook.NewLogger(factory)
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	sessionConfig := appConfig.Session
-	cookieDef := session.NewSessionCookieDef(request, sessionConfig, serverConfig)
-	redisLogger := redis3.NewLogger(factory)
-	redisStore := &redis3.Store{
-		Redis:  redisHandle,
-		AppID:  appID,
-		Clock:  clockClock,
-		Logger: redisLogger,
-	}
-	eventStore := &redis2.EventStore{
-		Redis: redisHandle,
-		AppID: appID,
-	}
-	accessEventProvider := &auth.AccessEventProvider{
-		Store: eventStore,
-	}
-	sessionRand := _wireRandValue
-	sessionProvider := &session.Provider{
-		Request:      request,
-		Store:        redisStore,
-		AccessEvents: accessEventProvider,
-		ServerConfig: serverConfig,
-		Config:       sessionConfig,
-		Clock:        clockClock,
-		Random:       sessionRand,
-	}
-	userController := &flows.UserController{
-		Users:         userProvider,
-		SessionCookie: cookieDef,
-		Sessions:      sessionProvider,
-		Hooks:         hookProvider,
-	}
-	webAppFlow := &flows.WebAppFlow{
-		Config:         identityConfig,
-		Identities:     providerProvider,
-		Users:          userProvider,
-		Hooks:          hookProvider,
-		Interactions:   interactionProvider,
-		UserController: userController,
-	}
-	oAuthClientCredentials := deps.ProvideOAuthClientCredentials(secretConfig)
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	userInfoDecoder := sso.UserInfoDecoder{
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	oAuthProviderFactory := &sso.OAuthProviderFactory{
-		Endpoints:                endpointsProvider,
-		IdentityConfig:           identityConfig,
-		Credentials:              oAuthClientCredentials,
-		RedirectURL:              urlProvider,
-		Clock:                    clockClock,
-		UserInfoDecoder:          userInfoDecoder,
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	authenticateProviderImpl := &webapp.AuthenticateProviderImpl{
-		ServerConfig:         serverConfig,
-		SSOOAuthConfig:       oAuthSSOConfig,
-		ValidateProvider:     validateProviderImpl,
-		RenderProvider:       renderProviderImpl,
-		StateProvider:        stateProviderImpl,
-		SSOStateCodec:        stateCodec,
-		Interactions:         webAppFlow,
-		OAuthProviderFactory: oAuthProviderFactory,
-	}
 	promoteHandler := &webapp2.PromoteHandler{
-		Provider: authenticateProviderImpl,
 		Database: handle,
 	}
 	return promoteHandler
@@ -2330,349 +1305,8 @@ func newWebAppPromoteHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppSSOCallbackHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	oAuthSSOConfig := identityConfig.OAuth
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	jwtKeyMaterials := deps.ProvideJWTKeyMaterials(secretConfig)
-	stateCodec := &sso.StateCodec{
-		AppID:       appID,
-		Clock:       clockClock,
-		Credentials: jwtKeyMaterials,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	messagingConfig := appConfig.Messaging
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	logger := hook.NewLogger(factory)
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	sessionConfig := appConfig.Session
-	cookieDef := session.NewSessionCookieDef(request, sessionConfig, serverConfig)
-	redisLogger := redis3.NewLogger(factory)
-	redisStore := &redis3.Store{
-		Redis:  redisHandle,
-		AppID:  appID,
-		Clock:  clockClock,
-		Logger: redisLogger,
-	}
-	eventStore := &redis2.EventStore{
-		Redis: redisHandle,
-		AppID: appID,
-	}
-	accessEventProvider := &auth.AccessEventProvider{
-		Store: eventStore,
-	}
-	sessionRand := _wireRandValue
-	sessionProvider := &session.Provider{
-		Request:      request,
-		Store:        redisStore,
-		AccessEvents: accessEventProvider,
-		ServerConfig: serverConfig,
-		Config:       sessionConfig,
-		Clock:        clockClock,
-		Random:       sessionRand,
-	}
-	userController := &flows.UserController{
-		Users:         userProvider,
-		SessionCookie: cookieDef,
-		Sessions:      sessionProvider,
-		Hooks:         hookProvider,
-	}
-	webAppFlow := &flows.WebAppFlow{
-		Config:         identityConfig,
-		Identities:     providerProvider,
-		Users:          userProvider,
-		Hooks:          hookProvider,
-		Interactions:   interactionProvider,
-		UserController: userController,
-	}
-	oAuthClientCredentials := deps.ProvideOAuthClientCredentials(secretConfig)
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	userInfoDecoder := sso.UserInfoDecoder{
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	oAuthProviderFactory := &sso.OAuthProviderFactory{
-		Endpoints:                endpointsProvider,
-		IdentityConfig:           identityConfig,
-		Credentials:              oAuthClientCredentials,
-		RedirectURL:              urlProvider,
-		Clock:                    clockClock,
-		UserInfoDecoder:          userInfoDecoder,
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	authenticateProviderImpl := &webapp.AuthenticateProviderImpl{
-		ServerConfig:         serverConfig,
-		SSOOAuthConfig:       oAuthSSOConfig,
-		ValidateProvider:     validateProviderImpl,
-		RenderProvider:       renderProviderImpl,
-		StateProvider:        stateProviderImpl,
-		SSOStateCodec:        stateCodec,
-		Interactions:         webAppFlow,
-		OAuthProviderFactory: oAuthProviderFactory,
-	}
 	ssoCallbackHandler := &webapp2.SSOCallbackHandler{
-		Provider: authenticateProviderImpl,
 		Database: handle,
 	}
 	return ssoCallbackHandler
@@ -2680,349 +1314,8 @@ func newWebAppSSOCallbackHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppEnterLoginIDHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	oAuthSSOConfig := identityConfig.OAuth
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	jwtKeyMaterials := deps.ProvideJWTKeyMaterials(secretConfig)
-	stateCodec := &sso.StateCodec{
-		AppID:       appID,
-		Clock:       clockClock,
-		Credentials: jwtKeyMaterials,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	messagingConfig := appConfig.Messaging
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	logger := hook.NewLogger(factory)
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	sessionConfig := appConfig.Session
-	cookieDef := session.NewSessionCookieDef(request, sessionConfig, serverConfig)
-	redisLogger := redis3.NewLogger(factory)
-	redisStore := &redis3.Store{
-		Redis:  redisHandle,
-		AppID:  appID,
-		Clock:  clockClock,
-		Logger: redisLogger,
-	}
-	eventStore := &redis2.EventStore{
-		Redis: redisHandle,
-		AppID: appID,
-	}
-	accessEventProvider := &auth.AccessEventProvider{
-		Store: eventStore,
-	}
-	sessionRand := _wireRandValue
-	sessionProvider := &session.Provider{
-		Request:      request,
-		Store:        redisStore,
-		AccessEvents: accessEventProvider,
-		ServerConfig: serverConfig,
-		Config:       sessionConfig,
-		Clock:        clockClock,
-		Random:       sessionRand,
-	}
-	userController := &flows.UserController{
-		Users:         userProvider,
-		SessionCookie: cookieDef,
-		Sessions:      sessionProvider,
-		Hooks:         hookProvider,
-	}
-	webAppFlow := &flows.WebAppFlow{
-		Config:         identityConfig,
-		Identities:     providerProvider,
-		Users:          userProvider,
-		Hooks:          hookProvider,
-		Interactions:   interactionProvider,
-		UserController: userController,
-	}
-	oAuthClientCredentials := deps.ProvideOAuthClientCredentials(secretConfig)
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	userInfoDecoder := sso.UserInfoDecoder{
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	oAuthProviderFactory := &sso.OAuthProviderFactory{
-		Endpoints:                endpointsProvider,
-		IdentityConfig:           identityConfig,
-		Credentials:              oAuthClientCredentials,
-		RedirectURL:              urlProvider,
-		Clock:                    clockClock,
-		UserInfoDecoder:          userInfoDecoder,
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	authenticateProviderImpl := &webapp.AuthenticateProviderImpl{
-		ServerConfig:         serverConfig,
-		SSOOAuthConfig:       oAuthSSOConfig,
-		ValidateProvider:     validateProviderImpl,
-		RenderProvider:       renderProviderImpl,
-		StateProvider:        stateProviderImpl,
-		SSOStateCodec:        stateCodec,
-		Interactions:         webAppFlow,
-		OAuthProviderFactory: oAuthProviderFactory,
-	}
 	enterLoginIDHandler := &webapp2.EnterLoginIDHandler{
-		Provider: authenticateProviderImpl,
 		Database: handle,
 	}
 	return enterLoginIDHandler
@@ -3030,349 +1323,8 @@ func newWebAppEnterLoginIDHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppEnterPasswordHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	oAuthSSOConfig := identityConfig.OAuth
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	jwtKeyMaterials := deps.ProvideJWTKeyMaterials(secretConfig)
-	stateCodec := &sso.StateCodec{
-		AppID:       appID,
-		Clock:       clockClock,
-		Credentials: jwtKeyMaterials,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	messagingConfig := appConfig.Messaging
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	logger := hook.NewLogger(factory)
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	sessionConfig := appConfig.Session
-	cookieDef := session.NewSessionCookieDef(request, sessionConfig, serverConfig)
-	redisLogger := redis3.NewLogger(factory)
-	redisStore := &redis3.Store{
-		Redis:  redisHandle,
-		AppID:  appID,
-		Clock:  clockClock,
-		Logger: redisLogger,
-	}
-	eventStore := &redis2.EventStore{
-		Redis: redisHandle,
-		AppID: appID,
-	}
-	accessEventProvider := &auth.AccessEventProvider{
-		Store: eventStore,
-	}
-	sessionRand := _wireRandValue
-	sessionProvider := &session.Provider{
-		Request:      request,
-		Store:        redisStore,
-		AccessEvents: accessEventProvider,
-		ServerConfig: serverConfig,
-		Config:       sessionConfig,
-		Clock:        clockClock,
-		Random:       sessionRand,
-	}
-	userController := &flows.UserController{
-		Users:         userProvider,
-		SessionCookie: cookieDef,
-		Sessions:      sessionProvider,
-		Hooks:         hookProvider,
-	}
-	webAppFlow := &flows.WebAppFlow{
-		Config:         identityConfig,
-		Identities:     providerProvider,
-		Users:          userProvider,
-		Hooks:          hookProvider,
-		Interactions:   interactionProvider,
-		UserController: userController,
-	}
-	oAuthClientCredentials := deps.ProvideOAuthClientCredentials(secretConfig)
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	userInfoDecoder := sso.UserInfoDecoder{
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	oAuthProviderFactory := &sso.OAuthProviderFactory{
-		Endpoints:                endpointsProvider,
-		IdentityConfig:           identityConfig,
-		Credentials:              oAuthClientCredentials,
-		RedirectURL:              urlProvider,
-		Clock:                    clockClock,
-		UserInfoDecoder:          userInfoDecoder,
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	authenticateProviderImpl := &webapp.AuthenticateProviderImpl{
-		ServerConfig:         serverConfig,
-		SSOOAuthConfig:       oAuthSSOConfig,
-		ValidateProvider:     validateProviderImpl,
-		RenderProvider:       renderProviderImpl,
-		StateProvider:        stateProviderImpl,
-		SSOStateCodec:        stateCodec,
-		Interactions:         webAppFlow,
-		OAuthProviderFactory: oAuthProviderFactory,
-	}
 	enterPasswordHandler := &webapp2.EnterPasswordHandler{
-		Provider: authenticateProviderImpl,
 		Database: handle,
 	}
 	return enterPasswordHandler
@@ -3380,349 +1332,8 @@ func newWebAppEnterPasswordHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppCreatePasswordHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	oAuthSSOConfig := identityConfig.OAuth
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	jwtKeyMaterials := deps.ProvideJWTKeyMaterials(secretConfig)
-	stateCodec := &sso.StateCodec{
-		AppID:       appID,
-		Clock:       clockClock,
-		Credentials: jwtKeyMaterials,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	messagingConfig := appConfig.Messaging
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	logger := hook.NewLogger(factory)
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	sessionConfig := appConfig.Session
-	cookieDef := session.NewSessionCookieDef(request, sessionConfig, serverConfig)
-	redisLogger := redis3.NewLogger(factory)
-	redisStore := &redis3.Store{
-		Redis:  redisHandle,
-		AppID:  appID,
-		Clock:  clockClock,
-		Logger: redisLogger,
-	}
-	eventStore := &redis2.EventStore{
-		Redis: redisHandle,
-		AppID: appID,
-	}
-	accessEventProvider := &auth.AccessEventProvider{
-		Store: eventStore,
-	}
-	sessionRand := _wireRandValue
-	sessionProvider := &session.Provider{
-		Request:      request,
-		Store:        redisStore,
-		AccessEvents: accessEventProvider,
-		ServerConfig: serverConfig,
-		Config:       sessionConfig,
-		Clock:        clockClock,
-		Random:       sessionRand,
-	}
-	userController := &flows.UserController{
-		Users:         userProvider,
-		SessionCookie: cookieDef,
-		Sessions:      sessionProvider,
-		Hooks:         hookProvider,
-	}
-	webAppFlow := &flows.WebAppFlow{
-		Config:         identityConfig,
-		Identities:     providerProvider,
-		Users:          userProvider,
-		Hooks:          hookProvider,
-		Interactions:   interactionProvider,
-		UserController: userController,
-	}
-	oAuthClientCredentials := deps.ProvideOAuthClientCredentials(secretConfig)
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	userInfoDecoder := sso.UserInfoDecoder{
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	oAuthProviderFactory := &sso.OAuthProviderFactory{
-		Endpoints:                endpointsProvider,
-		IdentityConfig:           identityConfig,
-		Credentials:              oAuthClientCredentials,
-		RedirectURL:              urlProvider,
-		Clock:                    clockClock,
-		UserInfoDecoder:          userInfoDecoder,
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	authenticateProviderImpl := &webapp.AuthenticateProviderImpl{
-		ServerConfig:         serverConfig,
-		SSOOAuthConfig:       oAuthSSOConfig,
-		ValidateProvider:     validateProviderImpl,
-		RenderProvider:       renderProviderImpl,
-		StateProvider:        stateProviderImpl,
-		SSOStateCodec:        stateCodec,
-		Interactions:         webAppFlow,
-		OAuthProviderFactory: oAuthProviderFactory,
-	}
 	createPasswordHandler := &webapp2.CreatePasswordHandler{
-		Provider: authenticateProviderImpl,
 		Database: handle,
 	}
 	return createPasswordHandler
@@ -3730,349 +1341,8 @@ func newWebAppCreatePasswordHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppOOBOTPHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	oAuthSSOConfig := identityConfig.OAuth
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	jwtKeyMaterials := deps.ProvideJWTKeyMaterials(secretConfig)
-	stateCodec := &sso.StateCodec{
-		AppID:       appID,
-		Clock:       clockClock,
-		Credentials: jwtKeyMaterials,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	messagingConfig := appConfig.Messaging
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	logger := hook.NewLogger(factory)
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	sessionConfig := appConfig.Session
-	cookieDef := session.NewSessionCookieDef(request, sessionConfig, serverConfig)
-	redisLogger := redis3.NewLogger(factory)
-	redisStore := &redis3.Store{
-		Redis:  redisHandle,
-		AppID:  appID,
-		Clock:  clockClock,
-		Logger: redisLogger,
-	}
-	eventStore := &redis2.EventStore{
-		Redis: redisHandle,
-		AppID: appID,
-	}
-	accessEventProvider := &auth.AccessEventProvider{
-		Store: eventStore,
-	}
-	sessionRand := _wireRandValue
-	sessionProvider := &session.Provider{
-		Request:      request,
-		Store:        redisStore,
-		AccessEvents: accessEventProvider,
-		ServerConfig: serverConfig,
-		Config:       sessionConfig,
-		Clock:        clockClock,
-		Random:       sessionRand,
-	}
-	userController := &flows.UserController{
-		Users:         userProvider,
-		SessionCookie: cookieDef,
-		Sessions:      sessionProvider,
-		Hooks:         hookProvider,
-	}
-	webAppFlow := &flows.WebAppFlow{
-		Config:         identityConfig,
-		Identities:     providerProvider,
-		Users:          userProvider,
-		Hooks:          hookProvider,
-		Interactions:   interactionProvider,
-		UserController: userController,
-	}
-	oAuthClientCredentials := deps.ProvideOAuthClientCredentials(secretConfig)
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	userInfoDecoder := sso.UserInfoDecoder{
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	oAuthProviderFactory := &sso.OAuthProviderFactory{
-		Endpoints:                endpointsProvider,
-		IdentityConfig:           identityConfig,
-		Credentials:              oAuthClientCredentials,
-		RedirectURL:              urlProvider,
-		Clock:                    clockClock,
-		UserInfoDecoder:          userInfoDecoder,
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	authenticateProviderImpl := &webapp.AuthenticateProviderImpl{
-		ServerConfig:         serverConfig,
-		SSOOAuthConfig:       oAuthSSOConfig,
-		ValidateProvider:     validateProviderImpl,
-		RenderProvider:       renderProviderImpl,
-		StateProvider:        stateProviderImpl,
-		SSOStateCodec:        stateCodec,
-		Interactions:         webAppFlow,
-		OAuthProviderFactory: oAuthProviderFactory,
-	}
 	oobotpHandler := &webapp2.OOBOTPHandler{
-		Provider: authenticateProviderImpl,
 		Database: handle,
 	}
 	return oobotpHandler
@@ -4080,309 +1350,8 @@ func newWebAppOOBOTPHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppForgotPasswordHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	messagingConfig := appConfig.Messaging
-	forgotPasswordConfig := appConfig.ForgotPassword
-	forgotpasswordStore := &forgotpassword.Store{
-		Redis: redisHandle,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	logger := hook.NewLogger(factory)
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	passwordFlow := &flows.PasswordFlow{
-		Interactions: interactionProvider,
-	}
-	forgotpasswordProvider := &forgotpassword.Provider{
-		Context:         context,
-		ServerConfig:    serverConfig,
-		Localization:    localizationConfig,
-		AppMetadata:     appMetadata,
-		Messaging:       messagingConfig,
-		Config:          forgotPasswordConfig,
-		Store:           forgotpasswordStore,
-		Users:           queries,
-		Hooks:           hookProvider,
-		Clock:           clockClock,
-		URLs:            urlProvider,
-		TemplateEngine:  engine,
-		TaskQueue:       queueQueue,
-		Interactions:    passwordFlow,
-		LoginIDProvider: loginidProvider,
-	}
-	forgotPasswordProvider := &webapp.ForgotPasswordProvider{
-		ValidateProvider: validateProviderImpl,
-		RenderProvider:   renderProviderImpl,
-		StateProvider:    stateProviderImpl,
-		ForgotPassword:   forgotpasswordProvider,
-	}
 	forgotPasswordHandler := &webapp2.ForgotPasswordHandler{
-		Provider: forgotPasswordProvider,
 		Database: handle,
 	}
 	return forgotPasswordHandler
@@ -4390,309 +1359,8 @@ func newWebAppForgotPasswordHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppForgotPasswordSuccessHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	messagingConfig := appConfig.Messaging
-	forgotPasswordConfig := appConfig.ForgotPassword
-	forgotpasswordStore := &forgotpassword.Store{
-		Redis: redisHandle,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	logger := hook.NewLogger(factory)
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	passwordFlow := &flows.PasswordFlow{
-		Interactions: interactionProvider,
-	}
-	forgotpasswordProvider := &forgotpassword.Provider{
-		Context:         context,
-		ServerConfig:    serverConfig,
-		Localization:    localizationConfig,
-		AppMetadata:     appMetadata,
-		Messaging:       messagingConfig,
-		Config:          forgotPasswordConfig,
-		Store:           forgotpasswordStore,
-		Users:           queries,
-		Hooks:           hookProvider,
-		Clock:           clockClock,
-		URLs:            urlProvider,
-		TemplateEngine:  engine,
-		TaskQueue:       queueQueue,
-		Interactions:    passwordFlow,
-		LoginIDProvider: loginidProvider,
-	}
-	forgotPasswordProvider := &webapp.ForgotPasswordProvider{
-		ValidateProvider: validateProviderImpl,
-		RenderProvider:   renderProviderImpl,
-		StateProvider:    stateProviderImpl,
-		ForgotPassword:   forgotpasswordProvider,
-	}
 	forgotPasswordSuccessHandler := &webapp2.ForgotPasswordSuccessHandler{
-		Provider: forgotPasswordProvider,
 		Database: handle,
 	}
 	return forgotPasswordSuccessHandler
@@ -4700,309 +1368,8 @@ func newWebAppForgotPasswordSuccessHandler(p *deps.RequestProvider) http.Handler
 
 func newWebAppResetPasswordHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	messagingConfig := appConfig.Messaging
-	forgotPasswordConfig := appConfig.ForgotPassword
-	forgotpasswordStore := &forgotpassword.Store{
-		Redis: redisHandle,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	logger := hook.NewLogger(factory)
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	passwordFlow := &flows.PasswordFlow{
-		Interactions: interactionProvider,
-	}
-	forgotpasswordProvider := &forgotpassword.Provider{
-		Context:         context,
-		ServerConfig:    serverConfig,
-		Localization:    localizationConfig,
-		AppMetadata:     appMetadata,
-		Messaging:       messagingConfig,
-		Config:          forgotPasswordConfig,
-		Store:           forgotpasswordStore,
-		Users:           queries,
-		Hooks:           hookProvider,
-		Clock:           clockClock,
-		URLs:            urlProvider,
-		TemplateEngine:  engine,
-		TaskQueue:       queueQueue,
-		Interactions:    passwordFlow,
-		LoginIDProvider: loginidProvider,
-	}
-	forgotPasswordProvider := &webapp.ForgotPasswordProvider{
-		ValidateProvider: validateProviderImpl,
-		RenderProvider:   renderProviderImpl,
-		StateProvider:    stateProviderImpl,
-		ForgotPassword:   forgotpasswordProvider,
-	}
 	resetPasswordHandler := &webapp2.ResetPasswordHandler{
-		Provider: forgotPasswordProvider,
 		Database: handle,
 	}
 	return resetPasswordHandler
@@ -5010,755 +1377,23 @@ func newWebAppResetPasswordHandler(p *deps.RequestProvider) http.Handler {
 
 func newWebAppResetPasswordSuccessHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	identityConfig := appConfig.Identity
-	loginIDConfig := identityConfig.LoginID
-	uiConfig := appConfig.UI
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	authenticationConfig := appConfig.Authentication
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	messagingConfig := appConfig.Messaging
-	forgotPasswordConfig := appConfig.ForgotPassword
-	forgotpasswordStore := &forgotpassword.Store{
-		Redis: redisHandle,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	logger := hook.NewLogger(factory)
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	passwordFlow := &flows.PasswordFlow{
-		Interactions: interactionProvider,
-	}
-	forgotpasswordProvider := &forgotpassword.Provider{
-		Context:         context,
-		ServerConfig:    serverConfig,
-		Localization:    localizationConfig,
-		AppMetadata:     appMetadata,
-		Messaging:       messagingConfig,
-		Config:          forgotPasswordConfig,
-		Store:           forgotpasswordStore,
-		Users:           queries,
-		Hooks:           hookProvider,
-		Clock:           clockClock,
-		URLs:            urlProvider,
-		TemplateEngine:  engine,
-		TaskQueue:       queueQueue,
-		Interactions:    passwordFlow,
-		LoginIDProvider: loginidProvider,
-	}
-	forgotPasswordProvider := &webapp.ForgotPasswordProvider{
-		ValidateProvider: validateProviderImpl,
-		RenderProvider:   renderProviderImpl,
-		StateProvider:    stateProviderImpl,
-		ForgotPassword:   forgotpasswordProvider,
-	}
 	resetPasswordSuccessHandler := &webapp2.ResetPasswordSuccessHandler{
-		Provider: forgotPasswordProvider,
 		Database: handle,
 	}
 	return resetPasswordSuccessHandler
 }
 
 func newWebAppSettingsHandler(p *deps.RequestProvider) http.Handler {
-	appProvider := p.AppProvider
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	authenticationConfig := appConfig.Authentication
-	uiConfig := appConfig.UI
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
-	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	identityConfig := appConfig.Identity
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	loginIDConfig := identityConfig.LoginID
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	settingsHandler := &webapp2.SettingsHandler{
-		RenderProvider: renderProviderImpl,
-	}
+	settingsHandler := &webapp2.SettingsHandler{}
 	return settingsHandler
 }
 
 func newWebAppSettingsIdentityHandler(p *deps.RequestProvider) http.Handler {
 	appProvider := p.AppProvider
-	rootProvider := appProvider.RootProvider
-	serverConfig := rootProvider.ServerConfig
-	config := appProvider.Config
-	appConfig := config.AppConfig
-	authenticationConfig := appConfig.Authentication
-	uiConfig := appConfig.UI
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
-	secretConfig := config.SecretConfig
-	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
-	appID := appConfig.ID
-	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
-	request := p.Request
-	context := deps.ProvideRequestContext(request)
 	handle := appProvider.Database
-	sqlExecutor := db.SQLExecutor{
-		Context:  context,
-		Database: handle,
-	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
-	identityConfig := appConfig.Identity
-	store := &loginid.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	loginIDConfig := identityConfig.LoginID
-	reservedNameChecker := rootProvider.ReservedNameChecker
-	typeCheckerFactory := &loginid.TypeCheckerFactory{
-		Config:              loginIDConfig,
-		ReservedNameChecker: reservedNameChecker,
-	}
-	loginidChecker := &loginid.Checker{
-		Config:             loginIDConfig,
-		TypeCheckerFactory: typeCheckerFactory,
-	}
-	normalizerFactory := &loginid.NormalizerFactory{
-		Config: loginIDConfig,
-	}
-	loginidProvider := &loginid.Provider{
-		Store:             store,
-		Config:            loginIDConfig,
-		Checker:           loginidChecker,
-		NormalizerFactory: normalizerFactory,
-	}
-	oauthStore := &oauth3.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	oauthProvider := &oauth3.Provider{
-		Store: oauthStore,
-		Clock: clockClock,
-	}
-	anonymousStore := &anonymous.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	anonymousProvider := &anonymous.Provider{
-		Store: anonymousStore,
-		Clock: clockClock,
-	}
-	providerProvider := &provider.Provider{
-		Authentication: authenticationConfig,
-		Identity:       identityConfig,
-		LoginID:        loginidProvider,
-		OAuth:          oauthProvider,
-		Anonymous:      anonymousProvider,
-	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	oAuthSSOConfig := identityConfig.OAuth
-	validateProviderImpl := &webapp.ValidateProviderImpl{
-		LoginID: loginIDConfig,
-		UI:      uiConfig,
-	}
-	redisHandle := appProvider.Redis
-	stateStoreImpl := &webapp.StateStoreImpl{
-		Redis: redisHandle,
-	}
-	factory := appProvider.LoggerFactory
-	stateProviderLogger := webapp.NewStateProviderLogger(factory)
-	stateProviderImpl := &webapp.StateProviderImpl{
-		StateStore: stateStoreImpl,
-		Logger:     stateProviderLogger,
-	}
-	jwtKeyMaterials := deps.ProvideJWTKeyMaterials(secretConfig)
-	stateCodec := &sso.StateCodec{
-		AppID:       appID,
-		Clock:       clockClock,
-		Credentials: jwtKeyMaterials,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	messagingConfig := appConfig.Messaging
-	welcomeMessageConfig := appConfig.WelcomeMessage
-	captureTaskContext := deps.ProvideCaptureTaskContext(config)
-	inMemoryExecutor := rootProvider.TaskExecutor
-	queueQueue := &queue.Queue{
-		Database:       handle,
-		CaptureContext: captureTaskContext,
-		Executor:       inMemoryExecutor,
-	}
-	welcomemessageProvider := &welcomemessage.Provider{
-		Context:               context,
-		LocalizationConfig:    localizationConfig,
-		MetadataConfiguration: appMetadata,
-		MessagingConfig:       messagingConfig,
-		WelcomeMessageConfig:  welcomeMessageConfig,
-		TemplateEngine:        engine,
-		TaskQueue:             queueQueue,
-	}
-	queries := &user.Queries{
-		Store:      userStore,
-		Identities: providerProvider,
-	}
-	rawCommands := &user.RawCommands{
-		Store:                  userStore,
-		Clock:                  clockClock,
-		WelcomeMessageProvider: welcomemessageProvider,
-		Queries:                queries,
-	}
-	logger := hook.NewLogger(factory)
-	rawProvider := &user.RawProvider{
-		RawCommands: rawCommands,
-		Queries:     queries,
-	}
-	hookStore := &hook.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	hookConfig := appConfig.Hook
-	webhookKeyMaterials := deps.ProvideWebhookKeyMaterials(secretConfig)
-	mutatorFactory := &hook.MutatorFactory{
-		Users: rawProvider,
-	}
-	syncHTTPClient := hook.NewSyncHTTPClient(hookConfig)
-	asyncHTTPClient := hook.NewAsyncHTTPClient()
-	deliverer := &hook.Deliverer{
-		Config:         hookConfig,
-		Secret:         webhookKeyMaterials,
-		Clock:          clockClock,
-		MutatorFactory: mutatorFactory,
-		SyncHTTP:       syncHTTPClient,
-		AsyncHTTP:      asyncHTTPClient,
-	}
-	hookProvider := &hook.Provider{
-		Context:   context,
-		Logger:    logger,
-		Database:  handle,
-		Clock:     clockClock,
-		Users:     rawProvider,
-		Store:     hookStore,
-		Deliverer: deliverer,
-	}
-	commands := &user.Commands{
-		Raw:   rawCommands,
-		Hooks: hookProvider,
-	}
-	userProvider := &user.Provider{
-		Commands: commands,
-		Queries:  queries,
-	}
-	interactionLogger := interaction.NewLogger(factory)
-	passwordStore := &password.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	passwordLogger := password.NewLogger(factory)
-	passwordProvider := &password.Provider{
-		Store:           passwordStore,
-		Config:          authenticatorPasswordConfig,
-		Clock:           clockClock,
-		Logger:          passwordLogger,
-		PasswordHistory: historyStore,
-		PasswordChecker: checker,
-	}
-	totpStore := &totp.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorTOTPConfig := authenticatorConfig.TOTP
-	totpProvider := &totp.Provider{
-		Store:  totpStore,
-		Config: authenticatorTOTPConfig,
-		Clock:  clockClock,
-	}
-	authenticatorOOBConfig := authenticatorConfig.OOB
-	oobStore := &oob.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	endpointsProvider := &endpoints.Provider{
-		Request: request,
-		Config:  serverConfig,
-	}
-	oobProvider := &oob.Provider{
-		Context:        context,
-		Localization:   localizationConfig,
-		AppMetadata:    appMetadata,
-		Messaging:      messagingConfig,
-		Config:         authenticatorOOBConfig,
-		Store:          oobStore,
-		TemplateEngine: engine,
-		Endpoints:      endpointsProvider,
-		TaskQueue:      queueQueue,
-		Clock:          clockClock,
-	}
-	bearertokenStore := &bearertoken.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorBearerTokenConfig := authenticatorConfig.BearerToken
-	bearertokenProvider := &bearertoken.Provider{
-		Store:  bearertokenStore,
-		Config: authenticatorBearerTokenConfig,
-		Clock:  clockClock,
-	}
-	recoverycodeStore := &recoverycode.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
-	authenticatorRecoveryCodeConfig := authenticatorConfig.RecoveryCode
-	recoverycodeProvider := &recoverycode.Provider{
-		Store:  recoverycodeStore,
-		Config: authenticatorRecoveryCodeConfig,
-		Clock:  clockClock,
-	}
-	provider3 := &provider2.Provider{
-		Password:     passwordProvider,
-		TOTP:         totpProvider,
-		OOBOTP:       oobProvider,
-		BearerToken:  bearertokenProvider,
-		RecoveryCode: recoverycodeProvider,
-	}
-	interactionProvider := &interaction.Provider{
-		Clock:         clockClock,
-		Logger:        interactionLogger,
-		Identity:      providerProvider,
-		Authenticator: provider3,
-		User:          userProvider,
-		OOB:           oobProvider,
-		Hooks:         hookProvider,
-		Config:        authenticationConfig,
-	}
-	sessionConfig := appConfig.Session
-	cookieDef := session.NewSessionCookieDef(request, sessionConfig, serverConfig)
-	redisLogger := redis3.NewLogger(factory)
-	redisStore := &redis3.Store{
-		Redis:  redisHandle,
-		AppID:  appID,
-		Clock:  clockClock,
-		Logger: redisLogger,
-	}
-	eventStore := &redis2.EventStore{
-		Redis: redisHandle,
-		AppID: appID,
-	}
-	accessEventProvider := &auth.AccessEventProvider{
-		Store: eventStore,
-	}
-	sessionRand := _wireRandValue
-	sessionProvider := &session.Provider{
-		Request:      request,
-		Store:        redisStore,
-		AccessEvents: accessEventProvider,
-		ServerConfig: serverConfig,
-		Config:       sessionConfig,
-		Clock:        clockClock,
-		Random:       sessionRand,
-	}
-	userController := &flows.UserController{
-		Users:         userProvider,
-		SessionCookie: cookieDef,
-		Sessions:      sessionProvider,
-		Hooks:         hookProvider,
-	}
-	webAppFlow := &flows.WebAppFlow{
-		Config:         identityConfig,
-		Identities:     providerProvider,
-		Users:          userProvider,
-		Hooks:          hookProvider,
-		Interactions:   interactionProvider,
-		UserController: userController,
-	}
-	oAuthClientCredentials := deps.ProvideOAuthClientCredentials(secretConfig)
-	challengeProvider := &challenge.Provider{
-		Redis: redisHandle,
-		AppID: appID,
-		Clock: clockClock,
-	}
-	anonymousFlow := &flows.AnonymousFlow{
-		Config:       authenticationConfig,
-		Interactions: interactionProvider,
-		Anonymous:    anonymousProvider,
-		Challenges:   challengeProvider,
-	}
-	urlProvider := &webapp.URLProvider{
-		Endpoints: endpointsProvider,
-		Anonymous: anonymousFlow,
-		States:    stateStoreImpl,
-	}
-	userInfoDecoder := sso.UserInfoDecoder{
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	oAuthProviderFactory := &sso.OAuthProviderFactory{
-		Endpoints:                endpointsProvider,
-		IdentityConfig:           identityConfig,
-		Credentials:              oAuthClientCredentials,
-		RedirectURL:              urlProvider,
-		Clock:                    clockClock,
-		UserInfoDecoder:          userInfoDecoder,
-		LoginIDNormalizerFactory: normalizerFactory,
-	}
-	authenticateProviderImpl := &webapp.AuthenticateProviderImpl{
-		ServerConfig:         serverConfig,
-		SSOOAuthConfig:       oAuthSSOConfig,
-		ValidateProvider:     validateProviderImpl,
-		RenderProvider:       renderProviderImpl,
-		StateProvider:        stateProviderImpl,
-		SSOStateCodec:        stateCodec,
-		Interactions:         webAppFlow,
-		OAuthProviderFactory: oAuthProviderFactory,
-	}
 	settingsIdentityHandler := &webapp2.SettingsIdentityHandler{
-		RenderProvider: renderProviderImpl,
-		Provider:       authenticateProviderImpl,
-		Database:       handle,
+		Database: handle,
 	}
 	return settingsIdentityHandler
 }
@@ -5768,17 +1403,9 @@ func newWebAppLogoutHandler(p *deps.RequestProvider) http.Handler {
 	rootProvider := appProvider.RootProvider
 	serverConfig := rootProvider.ServerConfig
 	config := appProvider.Config
-	appConfig := config.AppConfig
-	authenticationConfig := appConfig.Authentication
-	uiConfig := appConfig.UI
-	localizationConfig := appConfig.Localization
-	appMetadata := appConfig.Metadata
-	engine := appProvider.TemplateEngine
-	authenticatorConfig := appConfig.Authenticator
-	authenticatorPasswordConfig := authenticatorConfig.Password
-	clockClock := _wireSystemClockValue
 	secretConfig := config.SecretConfig
 	databaseCredentials := deps.ProvideDatabaseCredentials(secretConfig)
+	appConfig := config.AppConfig
 	appID := appConfig.ID
 	sqlBuilder := db.ProvideSQLBuilder(databaseCredentials, appID)
 	request := p.Request
@@ -5788,14 +1415,13 @@ func newWebAppLogoutHandler(p *deps.RequestProvider) http.Handler {
 		Context:  context,
 		Database: handle,
 	}
-	historyStore := &password.HistoryStore{
-		Clock:       clockClock,
+	store := &user.Store{
 		SQLBuilder:  sqlBuilder,
 		SQLExecutor: sqlExecutor,
 	}
-	checker := password.ProvideChecker(authenticatorPasswordConfig, historyStore)
+	authenticationConfig := appConfig.Authentication
 	identityConfig := appConfig.Identity
-	store := &loginid.Store{
+	loginidStore := &loginid.Store{
 		SQLBuilder:  sqlBuilder,
 		SQLExecutor: sqlExecutor,
 	}
@@ -5805,7 +1431,7 @@ func newWebAppLogoutHandler(p *deps.RequestProvider) http.Handler {
 		Config:              loginIDConfig,
 		ReservedNameChecker: reservedNameChecker,
 	}
-	loginidChecker := &loginid.Checker{
+	checker := &loginid.Checker{
 		Config:             loginIDConfig,
 		TypeCheckerFactory: typeCheckerFactory,
 	}
@@ -5813,15 +1439,16 @@ func newWebAppLogoutHandler(p *deps.RequestProvider) http.Handler {
 		Config: loginIDConfig,
 	}
 	loginidProvider := &loginid.Provider{
-		Store:             store,
+		Store:             loginidStore,
 		Config:            loginIDConfig,
-		Checker:           loginidChecker,
+		Checker:           checker,
 		NormalizerFactory: normalizerFactory,
 	}
 	oauthStore := &oauth3.Store{
 		SQLBuilder:  sqlBuilder,
 		SQLExecutor: sqlExecutor,
 	}
+	clockClock := _wireSystemClockValue
 	oauthProvider := &oauth3.Provider{
 		Store: oauthStore,
 		Clock: clockClock,
@@ -5841,28 +1468,17 @@ func newWebAppLogoutHandler(p *deps.RequestProvider) http.Handler {
 		OAuth:          oauthProvider,
 		Anonymous:      anonymousProvider,
 	}
-	renderProviderImpl := &webapp.RenderProviderImpl{
-		ServerConfig:                serverConfig,
-		AuthenticationConfiguration: authenticationConfig,
-		AuthUIConfiguration:         uiConfig,
-		LocalizationConfiguration:   localizationConfig,
-		MetadataConfiguration:       appMetadata,
-		TemplateEngine:              engine,
-		PasswordChecker:             checker,
-		Identity:                    providerProvider,
-	}
-	userStore := &user.Store{
-		SQLBuilder:  sqlBuilder,
-		SQLExecutor: sqlExecutor,
-	}
 	queries := &user.Queries{
-		Store:      userStore,
+		Store:      store,
 		Identities: providerProvider,
 	}
 	factory := appProvider.LoggerFactory
 	logger := hook.NewLogger(factory)
+	localizationConfig := appConfig.Localization
+	appMetadata := appConfig.Metadata
 	messagingConfig := appConfig.Messaging
 	welcomeMessageConfig := appConfig.WelcomeMessage
+	engine := appProvider.TemplateEngine
 	captureTaskContext := deps.ProvideCaptureTaskContext(config)
 	inMemoryExecutor := rootProvider.TaskExecutor
 	queueQueue := &queue.Queue{
@@ -5880,7 +1496,7 @@ func newWebAppLogoutHandler(p *deps.RequestProvider) http.Handler {
 		TaskQueue:             queueQueue,
 	}
 	rawCommands := &user.RawCommands{
-		Store:                  userStore,
+		Store:                  store,
 		Clock:                  clockClock,
 		WelcomeMessageProvider: welcomemessageProvider,
 		Queries:                queries,
@@ -5954,7 +1570,6 @@ func newWebAppLogoutHandler(p *deps.RequestProvider) http.Handler {
 	}
 	logoutHandler := &webapp2.LogoutHandler{
 		ServerConfig:   serverConfig,
-		RenderProvider: renderProviderImpl,
 		SessionManager: authSessionManager,
 		Database:       handle,
 	}
