+++ pkg/auth/dependency/webapp/deps.go
@@ -5,12 +5,7 @@ import (
 )
 
 var DependencySet = wire.NewSet(
-	wire.Struct(new(AuthenticateProviderImpl), "*"),
-	wire.Struct(new(ValidateProviderImpl), "*"),
-	wire.Bind(new(ValidateProvider), new(*ValidateProviderImpl)),
 	wire.Struct(new(ForgotPasswordProvider), "*"),
-	wire.Struct(new(RenderProviderImpl), "*"),
-	wire.Bind(new(RenderProvider), new(*RenderProviderImpl)),
 	wire.Struct(new(StateStoreImpl), "*"),
 	wire.Bind(new(StateStore), new(*StateStoreImpl)),
 	NewStateProviderLogger,
