+++ pkg/auth/handler/webapp/create_password.go
@@ -3,9 +3,66 @@ package webapp
 import (
 	"net/http"
 
+	"github.com/authgear/authgear-server/pkg/auth/config"
 	"github.com/authgear/authgear-server/pkg/db"
 	"github.com/authgear/authgear-server/pkg/httproute"
+	"github.com/authgear/authgear-server/pkg/template"
 )
+
+const (
+	// nolint: gosec
+	TemplateItemTypeAuthUICreatePasswordHTML config.TemplateItemType = "auth_ui_create_password.html"
+)
+
+var TemplateAuthUICreatePasswordHTML = template.Spec{
+	Type:        TemplateItemTypeAuthUICreatePasswordHTML,
+	IsHTML:      true,
+	Translation: TemplateItemTypeAuthUITranslationJSON,
+	Defines:     defines,
+	Components:  components,
+	Default: `<!DOCTYPE html>
+<html>
+{{ template "auth_ui_html_head.html" . }}
+<body class="page">
+<div class="content">
+
+{{ template "auth_ui_header.html" . }}
+
+<form class="simple-form vertical-form form-fields-container" method="post" novalidate>
+{{ $.csrfField }}
+
+<div class="nav-bar">
+	<button class="btn back-btn" type="button" title="{{ "back-button-title" }}"></button>
+	<div class="login-id primary-txt">
+	{{ if .x_national_number }}
+		+{{ .x_calling_code}} {{ .x_national_number }}
+	{{ else }}
+		{{ .x_login_id }}
+	{{ end }}
+	</div>
+</div>
+
+<div class="title primary-txt">{{ localize "create-password-page-title" }}</div>
+
+{{ template "ERROR" . }}
+
+<input id="password" data-password-policy-password="" class="input text-input primary-txt" type="password" name="x_password" placeholder="{{ localize "password-placeholder" }}">
+
+<button class="btn secondary-btn password-visibility-btn show-password" type="button">{{ localize "show-password" }}</button>
+<button class="btn secondary-btn password-visibility-btn hide-password" type="button">{{ localize "hide-password" }}</button>
+
+{{ template "PASSWORD_POLICY" . }}
+
+<button class="btn primary-btn align-self-flex-end" type="submit" name="submit" value="">{{ localize "next-button-label" }}</button>
+
+</form>
+{{ template "auth_ui_footer.html" . }}
+
+</div>
+</body>
+</html>
+`,
+}
 
 func ConfigureCreatePasswordRoute(route httproute.Route) httproute.Route {
 	return route.
@@ -13,13 +70,12 @@ func ConfigureCreatePasswordRoute(route httproute.Route) httproute.Route {
 		WithPathPattern("/create_password")
 }
 
-type CreatePasswordProvider interface {
-	GetCreatePasswordForm(w http.ResponseWriter, r *http.Request) (func(err error), error)
-	EnterSecret(w http.ResponseWriter, r *http.Request) (func(err error), error)
+type CreatePasswordViewModel struct {
+	BaseViewModel
+	PasswordPolicyViewModel
 }
 
 type CreatePasswordHandler struct {
-	Provider CreatePasswordProvider
 	Database *db.Handle
 }
 
@@ -30,17 +86,18 @@ func (h *CreatePasswordHandler) ServeHTTP(w http.ResponseWriter, r *http.Request
 	}
 
 	h.Database.WithTx(func() error {
-		if r.Method == "GET" {
-			writeResponse, err := h.Provider.GetCreatePasswordForm(w, r)
-			writeResponse(err)
-			return err
-		}
+		// FIXME(webapp): create_password
+		// if r.Method == "GET" {
+		// 	writeResponse, err := h.Provider.GetCreatePasswordForm(w, r)
+		// 	writeResponse(err)
+		// 	return err
+		// }
 
-		if r.Method == "POST" {
-			writeResponse, err := h.Provider.EnterSecret(w, r)
-			writeResponse(err)
-			return err
-		}
+		// if r.Method == "POST" {
+		// 	writeResponse, err := h.Provider.EnterSecret(w, r)
+		// 	writeResponse(err)
+		// 	return err
+		// }
 
 		return nil
 	})
