+++ pkg/auth/handler/webapp/enter_password.go
@@ -3,9 +3,82 @@ package webapp
 import (
 	"net/http"
 
+	"github.com/authgear/authgear-server/pkg/auth/config"
 	"github.com/authgear/authgear-server/pkg/db"
 	"github.com/authgear/authgear-server/pkg/httproute"
+	"github.com/authgear/authgear-server/pkg/template"
+	"github.com/authgear/authgear-server/pkg/validation"
 )
+
+const (
+	// nolint: gosec
+	TemplateItemTypeAuthUIEnterPasswordHTML config.TemplateItemType = "auth_ui_enter_password.html"
+)
+
+var TemplateAuthUIEnterPasswordHTML = template.Spec{
+	Type:        TemplateItemTypeAuthUIEnterPasswordHTML,
+	IsHTML:      true,
+	Translation: TemplateItemTypeAuthUITranslationJSON,
+	Defines:     defines,
+	Components:  components,
+	Default: `<!DOCTYPE html>
+<html>
+{{ template "auth_ui_html_head.html" . }}
+<body class="page">
+<div class="content">
+
+{{ template "auth_ui_header.html" . }}
+
+<form class="simple-form vertical-form form-fields-container" method="post" novalidate>
+{{ $.csrfField }}
+
+<div class="nav-bar">
+	<button class="btn back-btn" type="button" title="{{ localize "back-button-title" }}"></button>
+	<div class="login-id primary-txt">
+	{{ if .x_national_number }}
+		+{{ .x_calling_code}} {{ .x_national_number }}
+	{{ else }}
+		{{ .x_login_id }}
+	{{ end }}
+	</div>
+</div>
+
+<div class="title primary-txt">{{ localize "enter-password-page-title" }}</div>
+
+{{ template "ERROR" . }}
+
+<input id="password" class="input text-input primary-txt" type="password" name="x_password" placeholder="{{ localize "password-placeholder" }}">
+
+<button class="btn secondary-btn password-visibility-btn show-password" type="button">{{ localize "show-password" }}</button>
+<button class="btn secondary-btn password-visibility-btn hide-password" type="button">{{ localize "hide-password" }}</button>
+
+{{ if .x_password_authenticator_enabled }}
+<a class="link align-self-flex-start" href="{{ call .MakeURLWithPathWithoutX "/forgot_password" }}">{{ localize "forgot-password-button-label--enter-password-page" }}</a>
+{{ end }}
+
+<button class="btn primary-btn align-self-flex-end" type="submit" name="submit" value="">{{ localize "next-button-label" }}</button>
+
+</form>
+{{ template "auth_ui_footer.html" . }}
+
+</div>
+</body>
+</html>
+`,
+}
+
+const EnterPasswordRequest = "EnterPasswordRequest"
+
+var EnterPasswordSchema = validation.NewMultipartSchema("").
+	Add(EnterPasswordRequest, `
+		{
+			"type": "object",
+			"properties": {
+				"x_password": { "type": "string" }
+			},
+			"required": ["x_password"]
+		}
+	`).Instantiate()
 
 func ConfigureEnterPasswordRoute(route httproute.Route) httproute.Route {
 	return route.
@@ -13,13 +86,7 @@ func ConfigureEnterPasswordRoute(route httproute.Route) httproute.Route {
 		WithPathPattern("/enter_password")
 }
 
-type EnterPasswordProvider interface {
-	GetEnterPasswordForm(w http.ResponseWriter, r *http.Request) (func(err error), error)
-	EnterSecret(w http.ResponseWriter, r *http.Request) (func(err error), error)
-}
-
 type EnterPasswordHandler struct {
-	Provider EnterPasswordProvider
 	Database *db.Handle
 }
 
@@ -30,17 +97,19 @@ func (h *EnterPasswordHandler) ServeHTTP(w http.ResponseWriter, r *http.Request)
 	}
 
 	h.Database.WithTx(func() error {
-		if r.Method == "GET" {
-			writeResponse, err := h.Provider.GetEnterPasswordForm(w, r)
-			writeResponse(err)
-			return err
-		}
+		// FIXME(webapp): enter_password
+		// if r.Method == "GET" {
+		// 	writeResponse, err := h.Provider.GetEnterPasswordForm(w, r)
+		// 	writeResponse(err)
+		// 	return err
+		// }
 
-		if r.Method == "POST" {
-			writeResponse, err := h.Provider.EnterSecret(w, r)
-			writeResponse(err)
-			return err
-		}
+		// if r.Method == "POST" {
+		// 	writeResponse, err := h.Provider.EnterSecret(w, r)
+		// 	writeResponse(err)
+		// 	return err
+		// }
+
 		return nil
 	})
 }
