+++ pkg/auth/handler/webapp/forgot_password_success.go
@@ -3,9 +3,51 @@ package webapp
 import (
 	"net/http"
 
+	"github.com/authgear/authgear-server/pkg/auth/config"
 	"github.com/authgear/authgear-server/pkg/db"
 	"github.com/authgear/authgear-server/pkg/httproute"
+	"github.com/authgear/authgear-server/pkg/template"
 )
+
+const (
+	// nolint: gosec
+	TemplateItemTypeAuthUIForgotPasswordSuccessHTML config.TemplateItemType = "auth_ui_forgot_password_success.html"
+)
+
+var TemplateAuthUIForgotPasswordSuccessHTML = template.Spec{
+	Type:        TemplateItemTypeAuthUIForgotPasswordSuccessHTML,
+	IsHTML:      true,
+	Translation: TemplateItemTypeAuthUITranslationJSON,
+	Defines:     defines,
+	Components:  components,
+	Default: `<!DOCTYPE html>
+<html>
+{{ template "auth_ui_html_head.html" . }}
+<body class="page">
+<div class="content">
+
+{{ template "auth_ui_header.html" . }}
+
+<div class="simple-form vertical-form form-fields-container">
+
+<div class="title primary-txt">{{ localize "forgot-password-success-page-title" }}</div>
+
+{{ template "ERROR" . }}
+
+<!-- FIXME: x_login_id -->
+
+<div class="description primary-txt">{{ localize "forgot-password-success-description" .x_login_id }}</div>
+
+<a class="btn primary-btn align-self-flex-end" href="{{ call .MakeURLWithPathWithoutX "/login" }}">{{ localize "login-button-label--forgot-password-success-page" }}</a>
+
+</div>
+{{ template "auth_ui_footer.html" . }}
+
+</div>
+</body>
+</html>
+`,
+}
 
 func ConfigureForgotPasswordSuccessRoute(route httproute.Route) httproute.Route {
 	return route.
@@ -13,12 +55,7 @@ func ConfigureForgotPasswordSuccessRoute(route httproute.Route) httproute.Route 
 		WithPathPattern("/forgot_password/success")
 }
 
-type ForgotPasswordSuccessProvider interface {
-	GetForgotPasswordSuccess(w http.ResponseWriter, r *http.Request) (func(error), error)
-}
-
 type ForgotPasswordSuccessHandler struct {
-	Provider ForgotPasswordSuccessProvider
 	Database *db.Handle
 }
 
@@ -29,11 +66,13 @@ func (h *ForgotPasswordSuccessHandler) ServeHTTP(w http.ResponseWriter, r *http.
 	}
 
 	h.Database.WithTx(func() error {
-		if r.Method == "GET" {
-			writeResponse, err := h.Provider.GetForgotPasswordSuccess(w, r)
-			writeResponse(err)
-			return err
-		}
+		// FIXME(webapp): forgot_password_success
+		// if r.Method == "GET" {
+		// 	writeResponse, err := h.Provider.GetForgotPasswordSuccess(w, r)
+		// 	writeResponse(err)
+		// 	return err
+		// }
+
 		return nil
 	})
 }
