+++ pkg/auth/handler/webapp/logout.go
@@ -8,7 +8,40 @@ import (
 	"github.com/authgear/authgear-server/pkg/auth/dependency/webapp"
 	"github.com/authgear/authgear-server/pkg/db"
 	"github.com/authgear/authgear-server/pkg/httproute"
+	"github.com/authgear/authgear-server/pkg/template"
 )
+
+const (
+	TemplateItemTypeAuthUILogoutHTML config.TemplateItemType = "auth_ui_logout.html"
+)
+
+var TemplateAuthUILogoutHTML = template.Spec{
+	Type:        TemplateItemTypeAuthUILogoutHTML,
+	IsHTML:      true,
+	Translation: TemplateItemTypeAuthUITranslationJSON,
+	Defines:     defines,
+	Components:  components,
+	Default: `<!DOCTYPE html>
+<html>
+{{ template "auth_ui_html_head.html" . }}
+<body class="page">
+<div class="content">
+
+{{ template "auth_ui_header.html" . }}
+
+<form class="logout-form" method="post" novalidate>
+  {{ $.csrfField }}
+  <p class="primary-txt">{{ localize "logout-button-hint" }}</p>
+  <button class="btn primary-btn align-self-center" type="submit" name="x_action" value="logout">{{ localize "logout-button-label" }}</button>
+</form>
+
+{{ template "auth_ui_footer.html" . }}
+
+</div>
+</body>
+</html>
+`,
+}
 
 func ConfigureLogoutRoute(route httproute.Route) httproute.Route {
 	return route.
@@ -22,7 +55,6 @@ type LogoutSessionManager interface {
 
 type LogoutHandler struct {
 	ServerConfig   *config.ServerConfig
-	RenderProvider webapp.RenderProvider
 	SessionManager LogoutSessionManager
 	Database       *db.Handle
 }
@@ -34,7 +66,8 @@ func (h *LogoutHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 			h.SessionManager.Logout(sess, w)
 			webapp.RedirectToRedirectURI(w, r, h.ServerConfig.TrustProxy)
 		} else {
-			h.RenderProvider.WritePage(w, r, webapp.TemplateItemTypeAuthUILogoutHTML, nil)
+			// FIXME(webapp): logout
+			// h.RenderProvider.WritePage(w, r, webapp.TemplateItemTypeAuthUILogoutHTML, nil)
 		}
 		return nil
 	})
