+++ pkg/auth/handler/webapp/reset_password_success.go
@@ -3,9 +3,49 @@ package webapp
 import (
 	"net/http"
 
+	"github.com/authgear/authgear-server/pkg/auth/config"
 	"github.com/authgear/authgear-server/pkg/db"
 	"github.com/authgear/authgear-server/pkg/httproute"
+	"github.com/authgear/authgear-server/pkg/template"
 )
+
+const (
+	// nolint: gosec
+	TemplateItemTypeAuthUIResetPasswordSuccessHTML config.TemplateItemType = "auth_ui_reset_password_success.html"
+)
+
+var TemplateAuthUIResetPasswordSuccessHTML = template.Spec{
+	Type:        TemplateItemTypeAuthUIResetPasswordSuccessHTML,
+	IsHTML:      true,
+	Translation: TemplateItemTypeAuthUITranslationJSON,
+	Defines:     defines,
+	Components:  components,
+	Default: `<!DOCTYPE html>
+<html>
+{{ template "auth_ui_html_head.html" . }}
+<body class="page">
+<div class="content">
+
+{{ template "auth_ui_header.html" . }}
+
+<div class="simple-form vertical-form form-fields-container">
+
+<div class="title primary-txt">{{ localize "reset-password-success-page-title" }}</div>
+
+{{ template "ERROR" . }}
+
+<!-- FIXME: x_login_id -->
+
+<div class="description primary-txt">{{ localize "reset-password-success-description" .x_login_id }}</div>
+
+</div>
+{{ template "auth_ui_footer.html" . }}
+
+</div>
+</body>
+</html>
+`,
+}
 
 func ConfigureResetPasswordSuccessRoute(route httproute.Route) httproute.Route {
 	return route.
@@ -13,12 +53,7 @@ func ConfigureResetPasswordSuccessRoute(route httproute.Route) httproute.Route {
 		WithPathPattern("/reset_password/success")
 }
 
-type ResetPasswordSuccessProvider interface {
-	GetResetPasswordSuccess(w http.ResponseWriter, r *http.Request) (func(error), error)
-}
-
 type ResetPasswordSuccessHandler struct {
-	Provider ResetPasswordSuccessProvider
 	Database *db.Handle
 }
 
@@ -29,11 +64,12 @@ func (h *ResetPasswordSuccessHandler) ServeHTTP(w http.ResponseWriter, r *http.R
 	}
 
 	h.Database.WithTx(func() error {
-		if r.Method == "GET" {
-			writeResponse, err := h.Provider.GetResetPasswordSuccess(w, r)
-			writeResponse(err)
-			return err
-		}
+		// FIXME(webapp): reset_password_success
+		// if r.Method == "GET" {
+		// 	writeResponse, err := h.Provider.GetResetPasswordSuccess(w, r)
+		// 	writeResponse(err)
+		// 	return err
+		// }
 		return nil
 	})
 }
