+++ pkg/auth/handler/webapp/settings.go
@@ -3,9 +3,40 @@ package webapp
 import (
 	"net/http"
 
-	"github.com/authgear/authgear-server/pkg/auth/dependency/webapp"
+	"github.com/authgear/authgear-server/pkg/auth/config"
 	"github.com/authgear/authgear-server/pkg/httproute"
+	"github.com/authgear/authgear-server/pkg/template"
 )
+
+const (
+	TemplateItemTypeAuthUISettingsHTML config.TemplateItemType = "auth_ui_settings.html"
+)
+
+var TemplateAuthUISettingsHTML = template.Spec{
+	Type:        TemplateItemTypeAuthUISettingsHTML,
+	IsHTML:      true,
+	Translation: TemplateItemTypeAuthUITranslationJSON,
+	Defines:     defines,
+	Components:  components,
+	Default: `<!DOCTYPE html>
+<html>
+{{ template "auth_ui_html_head.html" . }}
+<body class="page">
+<div class="content">
+
+{{ template "auth_ui_header.html" . }}
+
+<div class="settings-form primary-txt">
+  You are authenticated. To logout, please visit <a href="/logout">here</a>.
+</div>
+
+{{ template "auth_ui_footer.html" . }}
+
+</div>
+</body>
+</html>
+`,
+}
 
 func ConfigureSettingsRoute(route httproute.Route) httproute.Route {
 	return route.
@@ -14,9 +45,10 @@ func ConfigureSettingsRoute(route httproute.Route) httproute.Route {
 }
 
 type SettingsHandler struct {
-	RenderProvider webapp.RenderProvider
+	// RenderProvider webapp.RenderProvider
 }
 
 func (h *SettingsHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	h.RenderProvider.WritePage(w, r, webapp.TemplateItemTypeAuthUISettingsHTML, nil)
+	// FIXME(webapp): settings
+	// h.RenderProvider.WritePage(w, r, webapp.TemplateItemTypeAuthUISettingsHTML, nil)
 }
