+++ pkg/auth/handler/webapp/sso_callback.go
@@ -13,12 +13,7 @@ func ConfigureSSOCallbackRoute(route httproute.Route) httproute.Route {
 		WithPathPattern("/sso/oauth2/callback/:alias")
 }
 
-type SSOProvider interface {
-	HandleSSOCallback(w http.ResponseWriter, r *http.Request, providerAlias string) (func(error), error)
-}
-
 type SSOCallbackHandler struct {
-	Provider SSOProvider
 	Database *db.Handle
 }
 
@@ -28,11 +23,14 @@ func (h *SSOCallbackHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	providerAlias := httproute.GetParam(r, "alias")
+	// providerAlias := httproute.GetParam(r, "alias")
 
 	h.Database.WithTx(func() error {
-		writeResponse, err := h.Provider.HandleSSOCallback(w, r, providerAlias)
-		writeResponse(err)
-		return err
+		// FIXME(webapp): sso_callback
+		// writeResponse, err := h.Provider.HandleSSOCallback(w, r, providerAlias)
+		// writeResponse(err)
+		// return err
+
+		return nil
 	})
 }
